+++ src/ui/react/settings/app.tsx
@@ -6,6 +6,7 @@ import { useSelector } from "react-redux";
 import { newSelectors } from "./modules/shared/store/app/reducer.ts";
 
 import { Routing } from "./router.tsx";
+import { ThumbnailGeneratorModal } from "./components/ThumbnailGeneratorModal/index.tsx";
 
 export function App() {
   const isDarkMode = useDarkMode();
@@ -30,6 +31,7 @@ export function App() {
       }}
     >
       <Routing />
+      <ThumbnailGeneratorModal />
     </ConfigProvider>
   );
 }